refactor: move CLI runtime behind MCP adapter
